Adjust admin keys to copy over existing root keys
When we make a digital ocean droplet, we add in the ssh keys we
want on that droplet.  These get added to the root user's authorized keys.

This hardening role will remove login access to root, letting you log in
as an admin user instead.  Before, we would take a key given by the
ansible role runner and put that in as the admin's key, but that was incorrect
because it would ignore cases where the droplet was designed for
multiple people to access it.

Now, we copy over the existing authorized keys from root to admin, maintaining
the correct access for our team, but no longer allowing access to root.