Add vaultpass script for vault decryptions
This makes it so that ansible-vault grabs your vault password from your
env vars, so you odn't have to enter it every time and you don't store
it on dosk somewhere to be accidentally committed.

What this doesn't cover is adding the pass to your env vars.  This is
done with export VAULT_PASS=pass, which can be included in your .envrc
if you're using direnv for additional convenience