Add inventory directory with encrypt vault vars
This sets up a new pattern where the inventory files are held
in their own directory, using group vars defined in the same dir as the
inventory file.  This makes working with the ansible vualt easy, while
losing some flexibility with how inventory files are defined, though we
maybe don't need that flexibility (e.g. a host being both a notification
service and a ssb pub, on the fly.)