Add playbooks dir for better structure/naming
At the moment, we have a number of yml files in the root of the repo
that either have verbose names (something-playbook.yml or
something-inventory.yml), or an unclear name (foo-prod.yml). This adds a
playbooks dir, in addition to the existing inventories dir, so that all
files can be namespaced by the dir and be named after the concern.
inventories/foo.yml and playbooks/foo.yml are intuitively paired.